+++ repository/resolver_test.go
@@ -1,4 +1,4 @@
-package main
+package repository
 
 import "testing"
